+++ src/main/java/com/sofodev/sworddisplay/data/DataGeneratorEvent.java
@@ -1,0 +1,18 @@
+package com.sofodev.sworddisplay.data;
+
+import net.minecraft.data.DataGenerator;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.event.lifecycle.GatherDataEvent;
+
+import static com.sofodev.sworddisplay.SwordDisplay.MODID;
+
+@Mod.EventBusSubscriber(modid = MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
+public class DataGeneratorEvent {
+
+    @SubscribeEvent
+    public static void gatherData(GatherDataEvent event) {
+        DataGenerator generator = event.getGenerator();
+        generator.addProvider(new Recipes(generator));
+    }
+}